# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "100kbetting \u2013 Core API",
    "description": "Ingest och exponering av odds, predictions och bets.",
    "contact": {
      "name": "100kbetting",
      "url": "https://example.com/"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/predictions": {
      "post": {
        "tags": [
          "predictions",
          "predictions"
        ],
        "summary": "Bulk upsert predictions",
        "operationId": "post_predictions_predictions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionsBulkIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "predictions",
          "predictions"
        ],
        "summary": "Lista predictions",
        "operationId": "get_predictions_predictions_get",
        "parameters": [
          {
            "name": "match_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Match Id"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "selection_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Selection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictionIn": {
        "properties": {
          "match_id": {
            "type": "string",
            "title": "Match Id"
          },
          "model_id": {
            "type": "string",
            "format": "uuid",
            "title": "Model Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "selection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selection Id"
          },
          "probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Probability"
          },
          "odds_fair": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Odds Fair"
          },
          "features": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "match_id",
          "model_id",
          "version",
          "selection_id",
          "probability"
        ],
        "title": "PredictionIn",
        "examples": [
          {
            "features": {
              "home_form": 0.65
            },
            "match_id": "m1",
            "model_id": "00000000-0000-0000-0000-000000000001",
            "odds_fair": 1.72,
            "probability": 0.58,
            "selection_id": "bea8671c-e889-4e3d-91d3-b407bc186408",
            "version": "1.0"
          }
        ]
      },
      "PredictionItem": {
        "properties": {
          "prediction_id": {
            "type": "string",
            "title": "Prediction Id"
          },
          "match_id": {
            "type": "string",
            "title": "Match Id"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "selection_id": {
            "type": "string",
            "title": "Selection Id"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "odds_fair": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Odds Fair"
          },
          "features": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features"
          },
          "predicted_add": {
            "type": "string",
            "title": "Predicted Add"
          }
        },
        "type": "object",
        "required": [
          "prediction_id",
          "match_id",
          "model_id",
          "version",
          "selection_id",
          "probability",
          "odds_fair",
          "features",
          "predicted_add"
        ],
        "title": "PredictionItem"
      },
      "PredictionsBulkIn": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PredictionIn"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PredictionsBulkIn"
      },
      "PredictionsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PredictionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "next_offset"
        ],
        "title": "PredictionsPage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}